requirements.txt
langchain
langchain-community
langgraph
langchain-ollama
langchain-groq
python-dotenv
wikipedia
ipykernel
ipython
duckduckgo-search